# update get_percentiles to take in an array