--- a/dataprofiler/profilers/float_column_profile.py
+++ b/dataprofiler/profilers/float_column_profile.py
@@ -120,9 +120,6 @@ class FloatColumn(NumericStatsMixin, BaseColumnPrimitiveTypeProfiler):
         Property for profile. Returns the profile of the column.
         :return:
         """
-        histogram_method = self.histogram_bin_method_names[0]
-        if self.histogram_selection is not None:
-            histogram_method = self.histogram_selection
 
         profile = dict(
             min=self.np_type_to_type(self.min),