Reduction in histogram time by refactoring get_percentiles:

TLDR;

--profile time--
Old method, optimized old method, new method
all bin methods: 8.37,7.80,6.78
auto only method: 7.18,6.81,6.66

--merge time--
Old method, optimized old method, new method
all bin methods: 6.61,5.17,0.46
auto only method: 1.56,0.84,0.39



===========================================================================

All tests are ran on a 23MB file with 1 col of each type plus a NAN col (index was accidentally included so extra int).
Old time, all binning:
```
profile time: 8.37269401550293
merge time: 6.606444835662842
Wrote profile results to profile.py.lprof
Timer unit: 1e-06 s

Total time: 6.46363 s
File: data-profiler/dataprofiler/profilers/numerical_column_stats.py
Function: _add_helper_merge_profile_histograms at line 100

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   100                                               @BaseColumnProfiler._timeit(name="histogram_and_quantiles")
   101                                               @profile
   102                                               def _add_helper_merge_profile_histograms(self, other1, other2):
   103                                                   """
   104                                                   Adds histogram of two profiles together
   105
   106                                                   :param other1: profile1 being added to self
   107                                                   :type other1: BaseColumnProfiler
   108                                                   :param other2: profile2 being added to self
   109                                                   :type other2: BaseColumnProfiler
   110                                                   :return: None
   111                                                   """
   112                                                   # get available bin methods and set to current
   113        28         90.0      3.2      0.0          bin_methods = list(set(other1.histogram_bin_method_names) &
   114        14         24.0      1.7      0.0                             set(other2.histogram_bin_method_names))
   115        14          8.0      0.6      0.0          if not bin_methods:
   116                                                       raise ValueError('Profiles have no overlapping bin methods and '
   117                                                                        'therefore cannot be added together.')
   118        14         16.0      1.1      0.0          elif other1.user_set_histogram_bin and other2.user_set_histogram_bin:
   119                                                       if other1.user_set_histogram_bin != other2.user_set_histogram_bin:
   120                                                           warnings.warn('User set histogram bin counts did not match. '
   121                                                                         'Choosing the larger bin count.')
   122                                                       self.user_set_histogram_bin = max(other1.user_set_histogram_bin,
   123                                                                                         other2.user_set_histogram_bin)
   124
   125                                                   # initial creation of the profiler creates all methods, but
   126                                                   # only the methods which intersect should exist.
   127        14         11.0      0.8      0.0          self.histogram_bin_method_names = bin_methods
   128        14         19.0      1.4      0.0          self.histogram_methods = dict()
   129       112         74.0      0.7      0.0          for method in self.histogram_bin_method_names:
   130        98        114.0      1.2      0.0              self.histogram_methods[method] = {
   131        98         43.0      0.4      0.0                  'total_loss': 0,
   132        98         42.0      0.4      0.0                  'current_loss': 0,
   133        98         55.0      0.6      0.0                  'histogram': {
   134        98         35.0      0.4      0.0                      'bin_counts': None,
   135        98         39.0      0.4      0.0                      'bin_edges': None
   136                                                           }
   137                                                       }
   138
   139       112        124.0      1.1      0.0          for i, method in enumerate(self.histogram_bin_method_names):
   140       294    1208478.0   4110.5     18.7              combined_values = other1._histogram_to_array(
   141       196    1128633.0   5758.3     17.5                  method) + other2._histogram_to_array(method)
   142       196    2023978.0  10326.4     31.3              bin_counts, bin_edges = self._get_histogram(
   143        98         52.0      0.5      0.0                  combined_values, method)
   144        98        353.0      3.6      0.0              self.histogram_methods[method]['histogram']['bin_counts'] = \
   145        98        112.0      1.1      0.0                  bin_counts
   146        98        130.0      1.3      0.0              self.histogram_methods[method]['histogram']['bin_edges'] = bin_edges
   147
   148                                                   # Select histogram: always choose first profile selected method
   149                                                   # Either both profiles have the same selection or you at least use one
   150                                                   # of the profiles selected method
   151        14         21.0      1.5      0.0          self.histogram_selection = other1.histogram_selection
   152        14    2101176.0 150084.0     32.5          self._get_quantiles()
```

Sped up old method using optimized old technique, all bins:
```
profile time: 7.8034539222717285
merge time: 5.167731046676636
Wrote profile results to profile.py.lprof
Timer unit: 1e-06 s

Total time: 5.02375 s
File: data-profiler/dataprofiler/profilers/numerical_column_stats.py
Function: _add_helper_merge_profile_histograms at line 100

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   100                                               @BaseColumnProfiler._timeit(name="histogram_and_quantiles")
   101                                               @profile
   102                                               def _add_helper_merge_profile_histograms(self, other1, other2):
   103                                                   """
   104                                                   Adds histogram of two profiles together
   105
   106                                                   :param other1: profile1 being added to self
   107                                                   :type other1: BaseColumnProfiler
   108                                                   :param other2: profile2 being added to self
   109                                                   :type other2: BaseColumnProfiler
   110                                                   :return: None
   111                                                   """
   112                                                   # get available bin methods and set to current
   113        28         92.0      3.3      0.0          bin_methods = list(set(other1.histogram_bin_method_names) &
   114        14         24.0      1.7      0.0                             set(other2.histogram_bin_method_names))
   115        14         12.0      0.9      0.0          if not bin_methods:
   116                                                       raise ValueError('Profiles have no overlapping bin methods and '
   117                                                                        'therefore cannot be added together.')
   118        14         19.0      1.4      0.0          elif other1.user_set_histogram_bin and other2.user_set_histogram_bin:
   119                                                       if other1.user_set_histogram_bin != other2.user_set_histogram_bin:
   120                                                           warnings.warn('User set histogram bin counts did not match. '
   121                                                                         'Choosing the larger bin count.')
   122                                                       self.user_set_histogram_bin = max(other1.user_set_histogram_bin,
   123                                                                                         other2.user_set_histogram_bin)
   124
   125                                                   # initial creation of the profiler creates all methods, but
   126                                                   # only the methods which intersect should exist.
   127        14         12.0      0.9      0.0          self.histogram_bin_method_names = bin_methods
   128        14         22.0      1.6      0.0          self.histogram_methods = dict()
   129       112         75.0      0.7      0.0          for method in self.histogram_bin_method_names:
   130        98        116.0      1.2      0.0              self.histogram_methods[method] = {
   131        98         46.0      0.5      0.0                  'total_loss': 0,
   132        98         35.0      0.4      0.0                  'current_loss': 0,
   133        98         56.0      0.6      0.0                  'histogram': {
   134        98         44.0      0.4      0.0                      'bin_counts': None,
   135        98         45.0      0.5      0.0                      'bin_edges': None
   136                                                           }
   137                                                       }
   138
   139       112        147.0      1.3      0.0          for i, method in enumerate(self.histogram_bin_method_names):
   140       294    1380331.0   4695.0     27.5              combined_values = other1._histogram_to_array(
   141       196    1266846.0   6463.5     25.2                  method) + other2._histogram_to_array(method)
   142       196    2322834.0  11851.2     46.2              bin_counts, bin_edges = self._get_histogram(
   143        98         63.0      0.6      0.0                  combined_values, method)
   144        98        530.0      5.4      0.0              self.histogram_methods[method]['histogram']['bin_counts'] = \
   145        98        157.0      1.6      0.0                  bin_counts
   146        98        142.0      1.4      0.0              self.histogram_methods[method]['histogram']['bin_edges'] = bin_edges
   147
   148                                                   # Select histogram: always choose first profile selected method
   149                                                   # Either both profiles have the same selection or you at least use one
   150                                                   # of the profiles selected method
   151        14         29.0      2.1      0.0          self.histogram_selection = other1.histogram_selection
   152        14      52076.0   3719.7      1.0          self._get_quantiles()
```

New method, all binning methods:
```
profile time: 6.7768378257751465
merge time: 0.4640200138092041
Wrote profile results to profile.py.lprof
Timer unit: 1e-06 s

Total time: 0.328324 s
File: data-profiler/dataprofiler/profilers/numerical_column_stats.py
Function: _add_helper_merge_profile_histograms at line 115

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   115                                               @BaseColumnProfiler._timeit(name="histogram_and_quantiles")
   116                                               @profile
   117                                               def _add_helper_merge_profile_histograms(self, other1, other2):
   118                                                   """
   119                                                   Adds histogram of two profiles together
   120
   121                                                   :param other1: profile1 being added to self
   122                                                   :type other1: BaseColumnProfiler
   123                                                   :param other2: profile2 being added to self
   124                                                   :type other2: BaseColumnProfiler
   125                                                   :return: None
   126                                                   """
   127                                                   # get available bin methods and set to current
   128        28         75.0      2.7      0.0          bin_methods = list(set(other1.histogram_bin_method_names) &
   129        14         26.0      1.9      0.0                             set(other2.histogram_bin_method_names))
   130        14          8.0      0.6      0.0          if not bin_methods:
   131                                                       raise ValueError('Profiles have no overlapping bin methods and '
   132                                                                        'therefore cannot be added together.')
   133        14         15.0      1.1      0.0          elif other1.user_set_histogram_bin and other2.user_set_histogram_bin:
   134                                                       if other1.user_set_histogram_bin != other2.user_set_histogram_bin:
   135                                                           warnings.warn('User set histogram bin counts did not match. '
   136                                                                         'Choosing the larger bin count.')
   137                                                       self.user_set_histogram_bin = max(other1.user_set_histogram_bin,
   138                                                                                         other2.user_set_histogram_bin)
   139
   140                                                   # initial creation of the profiler creates all methods, but
   141                                                   # only the methods which intersect should exist.
   142        14          9.0      0.6      0.0          self.histogram_bin_method_names = bin_methods
   143        14         20.0      1.4      0.0          self.histogram_methods = dict()
   144       112         61.0      0.5      0.0          for method in self.histogram_bin_method_names:
   145        98        116.0      1.2      0.0              self.histogram_methods[method] = {
   146        98         44.0      0.4      0.0                  'total_loss': 0,
   147        98         44.0      0.4      0.0                  'current_loss': 0,
   148        98         47.0      0.5      0.0                  'histogram': {
   149        98         34.0      0.3      0.0                      'bin_counts': None,
   150        98         41.0      0.4      0.0                      'bin_edges': None
   151                                                           }
   152                                                       }
   153
   154        28     101747.0   3633.8     31.0          combined_values = np.concatenate([other1._histogram_to_array(),
   155        14      77088.0   5506.3     23.5                                            other2._histogram_to_array()])
   156        14     145988.0  10427.7     44.5          bin_counts, bin_edges = self._get_histogram(combined_values)
   157        14         55.0      3.9      0.0          self._stored_histogram['histogram']['bin_counts'] = bin_counts
   158        14         21.0      1.5      0.0          self._stored_histogram['histogram']['bin_edges'] = bin_edges
   159
   160        14       2885.0    206.1      0.9          self._get_quantiles()
```

Old method, only `auto` binning:
```
profile time: 7.176765203475952
merge time: 1.5585100650787354
Wrote profile results to profile.py.lprof
Timer unit: 1e-06 s

Total time: 1.41826 s
File: data-profiler/dataprofiler/profilers/numerical_column_stats.py
Function: _add_helper_merge_profile_histograms at line 100

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   100                                               @BaseColumnProfiler._timeit(name="histogram_and_quantiles")
   101                                               @profile
   102                                               def _add_helper_merge_profile_histograms(self, other1, other2):
   103                                                   """
   104                                                   Adds histogram of two profiles together
   105
   106                                                   :param other1: profile1 being added to self
   107                                                   :type other1: BaseColumnProfiler
   108                                                   :param other2: profile2 being added to self
   109                                                   :type other2: BaseColumnProfiler
   110                                                   :return: None
   111                                                   """
   112                                                   # get available bin methods and set to current
   113        28         63.0      2.2      0.0          bin_methods = list(set(other1.histogram_bin_method_names) &
   114        14         21.0      1.5      0.0                             set(other2.histogram_bin_method_names))
   115        14          5.0      0.4      0.0          if not bin_methods:
   116                                                       raise ValueError('Profiles have no overlapping bin methods and '
   117                                                                        'therefore cannot be added together.')
   118        14         18.0      1.3      0.0          elif other1.user_set_histogram_bin and other2.user_set_histogram_bin:
   119                                                       if other1.user_set_histogram_bin != other2.user_set_histogram_bin:
   120                                                           warnings.warn('User set histogram bin counts did not match. '
   121                                                                         'Choosing the larger bin count.')
   122                                                       self.user_set_histogram_bin = max(other1.user_set_histogram_bin,
   123                                                                                         other2.user_set_histogram_bin)
   124
   125                                                   # initial creation of the profiler creates all methods, but
   126                                                   # only the methods which intersect should exist.
   127        14         11.0      0.8      0.0          self.histogram_bin_method_names = bin_methods
   128        14         21.0      1.5      0.0          self.histogram_methods = dict()
   129        28         22.0      0.8      0.0          for method in self.histogram_bin_method_names:
   130        14         19.0      1.4      0.0              self.histogram_methods[method] = {
   131        14          7.0      0.5      0.0                  'total_loss': 0,
   132        14          6.0      0.4      0.0                  'current_loss': 0,
   133        14          9.0      0.6      0.0                  'histogram': {
…