@JGSweets 2021-04-12
> The idea looks fine to me. However, with small number of values, it's not clear it improves from the current method. Can you run evaluation of the current and the new methods on these data
> 
> ```
> Normal data
> [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
> [1, 2, 3, 4, 5, 6, 6, 6, 6, 6]
> [1, 1, 1, 1, 1, 2, 3, 4, 5, 6]
> [1, 2, 3, 3, 3, 3, 3, 3, 4, 5]
> 
> Sparse data
> [1, 10, 100, 1, 10, 100, 1, 10, 100]
> [1, 1, 10, 10, 100, 100, 100, 100, 100]
> [1, 1, 1, 1, 1, 10, 10, 100, 100]
> [1, 1, 10, 10, 10, 10, 10, 100, 100]
> ```
> 
> (For each kind of data, we have data skewed to the right, left, central, and uniformly distributed.)

```
np.quantile:
[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
25: 3.25, 50: 5.5, 75: 7.75


[1, 2, 3, 4, 5, 6, 6, 6, 6, 6]
25: 3.25, 50: 5.5, 75: 6.0


[1, 1, 1, 1, 1, 2, 3, 4, 5, 6]
25: 1.0, 50: 1.5, 75: 3.75


[1, 2, 3, 3, 3, 3, 3, 3, 4, 5]
25: 3.0, 50: 3.0, 75: 3.0


[1, 10, 100, 1, 10, 100, 1, 10, 100]
25: 1.0, 50: 10.0, 75: 100.0


[1, 1, 10, 10, 100, 100, 100, 100, 100]
25: 10.0, 50: 100.0, 75: 100.0


[1, 1, 1, 1, 1, 10, 10, 100, 100]
25: 1.0, 50: 1.0, 75: 10.0


[1, 1, 10, 10, 10, 10, 10, 100, 100]
25: 10.0, 50: 10.0, 75: 10.0
```

```
CURRENT:
[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
25: 2.125, 50: 5.5, 75: 7.75


[1, 2, 3, 4, 5, 6, 6, 6, 6, 6]
25: 2.666666666666667, 50: 5.166666666666667, 75: 5.166666666666667


[1, 1, 1, 1, 1, 2, 3, 4, 5, 6]
25: 1.5, 50: 1.5, 75: 3.5


[1, 2, 3, 3, 3, 3, 3, 3, 4, 5]
25: 3.0, 50: 3.0, 75: 3.0


[1, 10, 100, 1, 10, 100, 1, 10, 100]
25: 9.25, 50: 9.25, 75: 83.5


[1, 1, 10, 10, 100, 100, 100, 100, 100]
25: 10.9, 50: 70.30000000000001, 75: 70.30000000000001


[1, 1, 1, 1, 1, 10, 10, 100, 100]
25: 5.125, 50: 5.125, 75: 9.25


[1, 1, 10, 10, 10, 10, 10, 100, 100]
25: 9.25, 50: 9.25, 75: 9.25
```

```
NEW:
[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
25: 2.875, 50: 5.5, 75: 8.125


[1, 2, 3, 4, 5, 6, 6, 6, 6, 6]
25: 3.083333333333333, 50: 5.166666666666667, 75: 5.583333333333334


[1, 1, 1, 1, 1, 2, 3, 4, 5, 6]
25: 1.5, 50: 2.0, 75: 4.500000000000001


[1, 2, 3, 3, 3, 3, 3, 3, 4, 5]
25: 3.0833333333333335, 50: 3.5, 75: 3.9166666666666665


[1, 10, 100, 1, 10, 100, 1, 10, 100]
25: 7.1875, 50: 13.375, 75: 87.62499999999997


[1, 1, 10, 10, 100, 100, 100, 100, 100]
25: 12.137500000000001, 50: 82.17999999999998, 75: 91.08999999999999


[1, 1, 1, 1, 1, 10, 10, 100, 100]
25: 4.7125, 50: 8.425, 75: 16.46875


[1, 1, 10, 10, 10, 10, 10, 100, 100]
25: 6.303571428571429, 50: 11.607142857142858, 75: 16.910714285714285
```